refactor(parser): extract PathConstraints into types/path_constraints
Plan 2, step 2 of N. Move `PathConstraints` + its small impl +
`extract_path_constraints` into
`crates/toolr-core/src/parser/types/path_constraints.rs` (87 lines).

`is_toolr_arg_call` stays in `types/mod.rs` but is promoted to
`pub(super)` so both `path_constraints.rs` (here) and the
not-yet-extracted `arg_metadata.rs` (step 5) can reach it via
`use super::is_toolr_arg_call;`.

`types/mod.rs` shrinks 1137 → 1059 lines. Public surface preserved:
`parser::types::{PathConstraints, extract_path_constraints}` resolve
via the new `pub use path_constraints::*;`.

Tests: 24 / 24 parser::types::tests still pass. `cargo check
--workspace` clean.
