+++ crates/toolr-core/src/parser/types/mod.rs
@@ -12,38 +12,16 @@
 //! pass over the file.
 
 mod imports;
+mod path_constraints;
 
 pub use imports::{SourcesImports, TypeImports};
+pub use path_constraints::{PathConstraints, extract_path_constraints};
 
 use ruff_python_ast::{Expr, ExprCall, StmtFunctionDef};
 use serde::{Deserialize, Serialize};
 
 use super::symbols::{ArgSectionTable, EnumTable, TypeAliasTable};
 use crate::manifest::{ArgMetadata, Argument, ArgumentKind, HelpSection};
-
-/// Filesystem constraints layered on top of a `Path`/`AbsolutePath`/
-/// `ResolvedPath` parameter, expressed via `arg(must_exist=True, ...)`
-/// inside `Annotated[Path, arg(...)]`. Top-level fields fold together
-/// — for example `must_be_file` implies `must_exist`.
-#[derive(Debug, Clone, Copy, Default, PartialEq, Eq, Serialize, Deserialize)]
-pub struct PathConstraints {
-    #[serde(default)]
-    pub must_exist: bool,
-    #[serde(default)]
-    pub must_be_file: bool,
-    #[serde(default)]
-    pub must_be_dir: bool,
-}
-
-impl PathConstraints {
-    pub fn is_empty(&self) -> bool {
-        !self.must_exist && !self.must_be_file && !self.must_be_dir
-    }
-    /// Whether any kind of disk check is required.
-    pub fn requires_existence(&self) -> bool {
-        self.must_exist || self.must_be_file || self.must_be_dir
-    }
-}
 
 /// Every annotation shape toolr recognises end-to-end.
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
@@ -478,59 +456,7 @@ fn literal_u32(expr: &Expr) -> Option<u32> {
     }
 }
 
-/// Walk an `Annotated[T, arg(...), ...]` annotation and extract
-/// `PathConstraints` from any `arg(...)` call inside it. Returns
-/// `None` if the annotation isn't `Annotated[...]` or carries no
-/// path-related arg() metadata.
-pub fn extract_path_constraints(annotation: &Expr) -> Option<PathConstraints> {
-    let Expr::Subscript(sub) = annotation else {
-        return None;
-    };
-    let head = match sub.value.as_ref() {
-        Expr::Name(n) => n.id.as_str(),
-        Expr::Attribute(a) => a.attr.as_str(),
-        _ => return None,
-    };
-    if head != "Annotated" {
-        return None;
-    }
-    let elts: Vec<&Expr> = match sub.slice.as_ref() {
-        Expr::Tuple(t) => t.elts.iter().collect(),
-        single => vec![single],
-    };
-    let mut constraints = PathConstraints::default();
-    let mut hit = false;
-    for elt in elts.iter().skip(1) {
-        let Expr::Call(call) = elt else { continue };
-        if !is_toolr_arg_call(call) {
-            continue;
-        }
-        for kw in &call.arguments.keywords {
-            let Some(name) = kw.arg.as_ref().map(|n| n.as_str()) else {
-                continue;
-            };
-            let Expr::BooleanLiteral(b) = &kw.value else { continue };
-            match name {
-                "must_exist" => {
-                    constraints.must_exist = b.value;
-                    hit = true;
-                }
-                "must_be_file" => {
-                    constraints.must_be_file = b.value;
-                    hit = true;
-                }
-                "must_be_dir" => {
-                    constraints.must_be_dir = b.value;
-                    hit = true;
-                }
-                _ => {}
-            }
-        }
-    }
-    if hit { Some(constraints) } else { None }
-}
-
-fn is_toolr_arg_call(call: &ExprCall) -> bool {
+pub(super) fn is_toolr_arg_call(call: &ExprCall) -> bool {
     match call.func.as_ref() {
         Expr::Name(n) => n.id.as_str() == "arg",
         Expr::Attribute(a) => a.attr.as_str() == "arg",
